Alpha 0.2.2
Started making Kaitlyn's abilities work, and added an game manager for storing gameplay events.

+++ The Vault/Assets/Scripts/ScriptableObjects/KaitlynSO.cs
@@ -46,7 +46,7 @@ public class KaitlynSO : ScriptableObject
     //stores time dilation power-ups
     public int TimeDilation;
     //stores squeeze through power-ups
-    public float SqueezeThrough;
+    public int SqueezeThrough;
     //stores dig power-ups
     public int Dig;
     //stores flight power-ups
